refactor: move code into session folder; turn firebase login ui into shortcode
