+++ session/loader.php
@@ -1,0 +1,6 @@
+<?php
+if ( ! defined( 'ABSPATH' ) ) { exit; }
+
+require_once( __DIR__ . '/menu.php' );
+require_once( __DIR__ . '/shortcodes.php' );
+require_once( __DIR__ . '/session-api.php' );
